renaming back the json_streamer
